Move App/Transport async task to ClientConnection class
