+++ src/Microsoft.Azure.SignalR/ServerConnections/ServiceConnectionWritter.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Threading.Tasks;
+using Microsoft.Azure.SignalR.Protocol;
+
+namespace Microsoft.Azure.SignalR
+{
+    public delegate Task ServiceConnectionWritter(ServiceMessage serviceMessage);
+}
